+++ helpers/Event.js
@@ -74,7 +74,7 @@ class Event {
     let K = this.K
     let mods = []
     //select messaging modules
-    Object.keys(K.modules).forEach((key) => {
+    Object.keys(K.modules).map((key) => {
       let mod = K.modules[key]
       if(true !== mod.event) return
       mods.push(mod)